refactor attach code to be more device agnostic, remove debug port options
